Complete Optics physical component ownership
