+++ docs/glossary.md
@@ -139,7 +139,7 @@ a channel detector and is not forced into an area-frame API.
 | Term | Project meaning |
 |---|---|
 | Backends namespace | `AdaptiveOpticsSim.Backends`: the canonical Julia owner of semantic array-backend selectors, concrete compute-device identity, and backend extension protocols for allocation, FFTs, reductions, random generation, kernel execution, and synchronization. It owns numerical execution mechanics, not physical model kernels, task scheduling, or CPU/GPU placement policy. `AdaptiveOpticsSim` exports the module itself but does not provide root compatibility bindings for its contents. |
-| Optics namespace | `AdaptiveOpticsSim.Optics`: the canonical Julia owner of apertures, telescopes, sources, optical-plane metadata and products, pupil and electric-field formation, general propagation, direct imaging, sampled OPD, physical NCPA, controllable optics, and reusable physical WFS components. Its foundation slice is implemented; controllable optics and WFS components complete in the next namespace gate. Atmosphere-coupled propagation, detector response, composed WFS estimation, calibration synthesis, and plant visibility or scheduling belong to their respective owners. |
+| Optics namespace | `AdaptiveOpticsSim.Optics`: the implemented canonical Julia owner of apertures, telescopes, sources, optical-plane metadata and products, pupil and electric-field formation, general propagation, direct imaging, sampled OPD, physical NCPA, controllable optics, and reusable physical WFS components. Atmosphere-coupled propagation, detector response, composed WFS estimation, calibration synthesis, and plant visibility or scheduling belong to their respective owners. |
 | Array backend | A semantic software/provider family selected by `Backends.CPUBackend()`, `Backends.CUDABackend()`, `Backends.AMDGPUBackend()`, or `Backends.MetalBackend()`. It determines compatible array construction and numerical dispatch, but it does not identify one concrete accelerator when a process can address several devices in the same family. |
 | Compute device | The concrete host or accelerator on which array storage resides and prepared numerical work executes, represented by `Backends.HostComputeDevice` or a family-qualified `Backends.AcceleratorComputeDevice`. `Backends.compute_device(array)` reports this run-local identity; supported array views report their parent's identity. A compute device is distinct from an AO component being simulated, an optical-plane location, a worker/task, and a future placement resource record. A backend runtime's device ordinal is sufficient for same-process compatibility but is not a stable declared plant identity. |
 | Execution owner | One stable run-local logical owner prepared for exactly one path execution group or compatible device batch and its mutable workspace, backend, and compute device. It has exclusive write ownership, fixed group ordinals, one bounded due-work SPSC path, and one owner-specific bounded completion path. A deterministic policy may service that ownership synchronously; a threaded policy may bind it to one reusable task. `ExecutionOwnerID` is a positive run-local ordinal, not a declared optical identity, Julia task/thread affinity, compute device, RTC endpoint, transport role, or automatic placement decision. |
