+++ src/plant/plant.jl
@@ -37,6 +37,9 @@ import ..Backends:
 
 import ..Optics:
     AbstractCombinationPolicy,
+    AbstractDMActuatorModel,
+    AbstractDMInfluenceModel,
+    AbstractDMTopology,
     AbstractOpticalNormalization,
     AbstractOpticalPlaneKind,
     AbstractOpticalProduct,
@@ -46,24 +49,33 @@ import ..Optics:
     AbstractSpectralCoordinate,
     AbstractTelescope,
     AchromaticSpectralCoordinate,
+    ActuatorGridTopology,
     Asterism,
     CellIntegratedMeasure,
+    CircularModulation,
     CoherentFieldCombination,
     DMAdditive,
     DMApplyMode,
     DMReplace,
+    DeformableMirror,
+    DeformableMirrorParams,
+    DeformableMirrorState,
     DetectorPlane,
     DimensionlessNormalization,
     ElectricField,
     ExpandedSourceComposition,
     ExtendedSource,
     FocalPlane,
+    GaussianInfluenceWidth,
+    GaussianMechanicalCoupling,
     IncoherentIntensityAddition,
     IntegratedSpectralChannel,
     IntensityMap,
     LGSSource,
     LeafSourceComposition,
+    LinearStaticActuators,
     MetricCoordinates,
+    Misregistration,
     MonochromaticChannel,
     NCPA,
     NonCombinableProduct,
@@ -76,6 +88,7 @@ import ..Optics:
     PreparedBundledDirectImaging,
     PreparedDirectImaging,
     PreparedDirectImagingBatch,
+    PreparedFocalPlaneModulation,
     PreparedIncoherentDirectImaging,
     PupilFunction,
     PupilPlane,
@@ -89,40 +102,41 @@ import ..Optics:
     UnspecifiedSpectralCoordinate,
     _FixedOpticalProductVector,
     _direct_imaging_batch_product_contract,
+    actuator_model,
     aperture_revision,
     coordinates_xy_arcsec,
     direct_imaging_output,
     form_direct_image!,
     freeze_source,
+    influence_model,
+    n_actuators,
     prepare_direct_imaging_batch,
     pupil_reflectivity,
+    set_command!,
     source_composition_style,
     source_height_m,
     source_radiometric_value,
     source_radiometry,
     surface_opd,
+    topology,
+    topology_command_count,
+    update_cycle_averaged_circular_modulation!,
+    update_surface!,
     validate_direct_imaging_batch,
+    validate_dm_actuator_model,
     validate_plane_storage,
     wavelength
 
 import ..AdaptiveOpticsSim:
     AbstractAtmosphere,
-    AbstractDMActuatorModel,
-    AbstractDMInfluenceModel,
-    AbstractDMTopology,
     AbstractEMCCDAcquisitionMode,
     AbstractFrameTimingModel,
     AbstractTimedAtmosphere,
     AcquiredObservationPath,
-    ActuatorGridTopology,
     AdaptiveOpticsSimError,
     AtmosphereEpoch,
     AtmosphereLayerID,
     CMOSSensor,
-    CircularModulation,
-    DeformableMirror,
-    DeformableMirrorParams,
-    DeformableMirrorState,
     Detector,
     DetectorAcquisitionPlan,
     DirectMeasurementPath,
@@ -130,21 +144,16 @@ import ..AdaptiveOpticsSim:
     FrameReadoutProducts,
     FrameSensorType,
     FrameTransferAcquisition,
-    GaussianInfluenceWidth,
-    GaussianMechanicalCoupling,
     GlobalResetExposure,
     HgCdTeAvalancheArraySensor,
     InfiniteAtmosphereLayer,
     InfiniteMultiLayerAtmosphere,
     InvalidConfiguration,
-    LinearStaticActuators,
-    Misregistration,
     MovingAtmosphereLayer,
     MultiLayerAtmosphere,
     PreparedAtmosphereDirectionBatch,
     PreparedBioEdgeOpticalBundleFormation,
     PreparedBioEdgeOpticalFormation,
-    PreparedFocalPlaneModulation,
     PreparedPyramidOpticalBundleFormation,
     PreparedPyramidOpticalFormation,
     PreparedShackHartmannOpticalBundleFormation,
@@ -165,7 +174,6 @@ import ..AdaptiveOpticsSim:
     _validate_epoch_identity,
     accumulate_incremental_charge_generation!,
     acquire_wfs_observation!,
-    actuator_model,
     advance_by!,
     advance_thermal!,
     advance_to!,
@@ -186,12 +194,10 @@ import ..AdaptiveOpticsSim:
     finalize_incremental_capture!,
     finalize_scheduled_up_the_ramp_readout_products!,
     form_wfs_optical_products!,
-    influence_model,
     initialize_atmosphere!,
     is_global_shutter,
     line_time,
     measurement_storage,
-    n_actuators,
     output_frame,
     prepare_atmosphere_direction_batch,
     prepare_atmosphere_renderer,
@@ -203,17 +209,11 @@ import ..AdaptiveOpticsSim:
     runtime_rng,
     sample_frame_read!,
     sampling_read_time,
-    set_command!,
     splitmix64,
     subtract_background_map!,
-    topology,
-    topology_command_count,
-    update_cycle_averaged_circular_modulation!,
     update_sensor_persistence!,
-    update_surface!,
     validate_atmosphere_direction_batch,
     validate_atmosphere_rendering,
-    validate_dm_actuator_model,
     validate_wfs_acquisition_binding,
     validate_wfs_estimation_binding,
     validate_wfs_measurement,
