+++ test/contracts/namespace_migration_state.toml
@@ -2,163 +2,13 @@ schema_version = 1
 name = "namespace_migration_state"
 status = "maintained"
 authority = "namespace_authority.toml"
-current_gate = "NS-02"
-implemented_owners = ["Backends"]
-partial_owners = ["Optics"]
+current_gate = "NS-03"
+implemented_owners = ["Backends", "Optics"]
+partial_owners = []
 
 [owner_sources]
 Backends = "src/backends/api.jl"
-
-[partial_owner_sources]
 Optics = "src/optics/api.jl"
-
-[partial_owner_bindings.Optics]
-exports = [
-    "CircularAperture",
-    "AnnularAperture",
-    "SpiderMask",
-    "RectangularROI",
-    "SubapertureGridMask",
-    "build_mask!",
-    "apply_mask!",
-    "Telescope",
-    "Source",
-    "LGSSource",
-    "Asterism",
-    "AbstractSourceRadiometry",
-    "PhysicalPhotonIrradianceSource",
-    "NormalizedTestSource",
-    "source_radiometry",
-    "source_radiometric_value",
-    "wavelength",
-    "photon_irradiance",
-    "optical_path",
-    "pupil_photon_rate_map",
-    "reset_opd!",
-    "apply_opd!",
-    "set_pupil!",
-    "set_pupil_reflectivity!",
-    "pupil_mask",
-    "pupil_reflectivity",
-    "apply_spiders!",
-    "focal_plane_pixel_scale_arcsec",
-    "SpectralSample",
-    "SpectralBundle",
-    "SpectralSource",
-    "with_spectrum",
-    "GaussianDiskSourceModel",
-    "PointCloudSourceModel",
-    "SampledImageSourceModel",
-    "with_extended_source",
-    "extended_source_asterism",
-    "AbstractOpticalProduct",
-    "AbstractOpticalPlaneKind",
-    "PupilPlane",
-    "FocalPlane",
-    "IntermediatePlane",
-    "DetectorPlane",
-    "AbstractPlaneCoordinateDomain",
-    "MetricCoordinates",
-    "AngularCoordinates",
-    "NormalizedPupilCoordinates",
-    "PlaneCentering",
-    "SampleCentered",
-    "InterSampleCentered",
-    "PlaneAxisOrientation",
-    "UnspecifiedSpectralCoordinate",
-    "AchromaticSpectralCoordinate",
-    "MonochromaticChannel",
-    "IntegratedSpectralChannel",
-    "AbstractOpticalNormalization",
-    "PhotonRateNormalization",
-    "DimensionlessNormalization",
-    "AbstractSpatialMeasure",
-    "PointSampledMeasure",
-    "SpatialDensityMeasure",
-    "CellIntegratedMeasure",
-    "AbstractCombinationPolicy",
-    "CoherentFieldCombination",
-    "IncoherentIntensityAddition",
-    "NonCombinableProduct",
-    "UnspecifiedNormalization",
-    "UnspecifiedSpatialMeasure",
-    "UnspecifiedCoherence",
-    "OpticalPlaneMetadata",
-    "plane_metadata",
-    "PupilFunction",
-    "pupil_support",
-    "pupil_amplitude",
-    "opd_map",
-    "surface_opd",
-    "apply_surface!",
-    "ElectricField",
-    "field_values",
-    "IntensityMap",
-    "intensity_values",
-    "OpticalProductBundle",
-    "PreparedIncoherentSum",
-    "prepare_incoherent_sum",
-    "accumulate_intensity!",
-    "PupilFieldFormationPlan",
-    "prepare_pupil_field",
-    "fill_electric_field!",
-    "fill_electric_field_async!",
-    "FraunhoferPropagation",
-    "FresnelPropagation",
-    "propagation_output",
-    "propagate_field!",
-    "fraunhofer_intensity_from_field!",
-    "DirectImagingPlan",
-    "DirectImagingWorkspace",
-    "prepare_direct_imaging",
-    "form_direct_image!",
-    "direct_imaging_output",
-    "direct_imaging_components",
-    "prepare_direct_imaging_batch",
-    "OPDMap",
-    "NCPA",
-    "DMAdditive",
-    "DMReplace",
-]
-public = [
-    "AbstractSource",
-    "AbstractDirectImagingBatchCapability",
-    "StackedFraunhoferDirectImagingBatchCapability",
-    "UnsupportedDirectImagingBatchCapability",
-    "DirectImagingBatchSampleParams",
-    "DirectImagingBatchProductContract",
-    "DirectImagingBatchCompatibilitySignature",
-    "DirectImagingBatchWorkspace",
-    "PreparedDirectImagingBatch",
-    "direct_imaging_batch_capability",
-    "direct_imaging_batch_signature",
-    "direct_imaging_batch_sources",
-    "direct_imaging_batch_inputs",
-    "direct_imaging_batch_products",
-    "direct_imaging_batch_count",
-    "validate_direct_imaging_batch",
-]
-internal = [
-    "AbstractOpticalElement",
-    "AbstractSourceCompositionStyle",
-    "AbstractSpectralCoordinate",
-    "AbstractTelescope",
-    "DMApplyMode",
-    "ExpandedSourceComposition",
-    "ExtendedSource",
-    "LeafSourceComposition",
-    "PreparedBundledDirectImaging",
-    "PreparedDirectImaging",
-    "PreparedIncoherentDirectImaging",
-    "_FixedOpticalProductVector",
-    "_direct_imaging_batch_product_contract",
-    "aperture_revision",
-    "coordinates_xy_arcsec",
-    "freeze_source",
-    "source_composition_style",
-    "source_height_m",
-    "validate_plane_storage",
-]
 
 [baseline_divergence]
 root_surface = "src/exports.jl"
